chore(RHINENG-16141): Cleanup CreatePolicy components
